google_drive_api/requirements.txt
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-auth==2.25.2
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
requests==2.31.0
flask==3.0.0
watchdog==3.0.0

